update native iOS dependency install
